+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # farmer change log
 
+## Unreleased
+
+### Added
+
+* `logdna`: Added `logdna` command to export lines from [LogDNA](https://logdna.com/).
+
 ## 2.3.0 (2018-01-04)
 
 ### Added